// payload-types.ts
export interface Tour {
  id: number;
  spot: string;
  /**
   * URL path, e.g. otgontenger -> /tour-details/otgontenger
   */
  slug: string;
  country?: string | null;
  image: number | Media;
  gallery?:
    | {
        image?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  price: number;
  discount?: number | null;
  day: number;
  night: number;
  /**
   * Number of countries covered
   */
  totalCountry?: number | null;
  shortDescription?: string | null;
  description?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  highlights?:
    | {
        text?: string | null;
        id?: string | null;
      }[]
    | null;
  included?:
    | {
        text?: string | null;
        id?: string | null;
      }[]
    | null;
  excluded?:
    | {
        text?: string | null;
        id?: string | null;
      }[]
    | null;
  itinerary?:
    | {
        title?: string | null;
        details?: string | null;
        id?: string | null;
      }[]
    | null;
  featured?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    feature?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}